+++ mitra_models/migrations/schema.sql
@@ -116,7 +116,7 @@ CREATE TABLE oauth_authorization (
     code VARCHAR(100) UNIQUE NOT NULL,
     user_id UUID NOT NULL REFERENCES user_account (id) ON DELETE CASCADE,
     application_id INTEGER NOT NULL REFERENCES oauth_application (id) ON DELETE CASCADE,
-    scopes VARCHAR(200) NOT NULL,
+    scopes TEXT[] NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     expires_at TIMESTAMP WITH TIME ZONE NOT NULL
 );
@@ -125,6 +125,7 @@ CREATE TABLE oauth_token (
     id SERIAL PRIMARY KEY,
     owner_id UUID NOT NULL REFERENCES user_account (id) ON DELETE CASCADE,
     application_id INTEGER REFERENCES oauth_application (id) ON DELETE CASCADE,
+    scopes TEXT[] NOT NULL,
     token_digest BYTEA UNIQUE NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
     expires_at TIMESTAMP WITH TIME ZONE NOT NULL