+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed default value for `exclude_replies` parameter to `false`.
 - Write warning to log if client ID is not provided with OAuth token request.
 - Write warning to log if unsupported OAuth scopes are requested.
+- Return scopes that were requested by the user in `/oauth/token` response.
 
 ### Fixed